# avoid 'inv' array in bin_sort