Unless I'm missing something, it should be possible to perform the bin sorting without the additional `inv` vector. The resulting code is shorter and should also be faster, since it avoids the writing and subsequent reading of `inv`, and the times for random memory accesses may be partially overlapped with the bin index computations.
I didn't do any benchmarking however ...