@ahbarnett 2023-06-20
Some tests on Ryzen 7 5700U laptop (8 cores). One easy sorting-dominated test is to go to 1D, low acc, large N.
By default dir=2 (interp) has sorting switched off, so we're only looking at the dir=1 times:
```
perftest/spreadtestnd 1 1e8 1e8 1e-1
nthreads      1     4   8   16 (full HT)
master         13.7s  4.6s   3.5s  6.6s        (notice slow-down with full HT)
PR#291       14s     4.3s    3.0s   3.6s
```
So, improvement at large thread count (esp 80% faster at full HT); no improvement for single-thread.
Now for single-prec:
```
perftest/spreadtestndf 1 1e8 1e8 1e-1
nthreads      1     4   8   16 (full HT)
master         10.7s  3.7s   3.5s  6.1s        (notice slow-down with full HT)
PR#291       10.6s     3.4s    2.4s   3.0s    (only slight slow-down with HT)
```
similar story. Very nice.