@ahbarnett 2023-06-20
We test xeon 16-core 6244 gold. Same test as for laptop above.
```
perftest/spreadtestnd 1 1e8 1e8 1e-1
nthreads      1    16      32 (full HT)
master:      9.8s   3.4s   4.9s
PR#291:     9.9s    2.15s  4.7s
```
We see >50% speedup at 1thread/core, but both suffer badly with full HT.
```
perftest/spreadtestndf 1 1e8 1e8 1e-1
nthreads       1    16    32
master:       9.6s   3.25s  4.5s
PR#291:      9.5s   2.12s  4.4s
```
Similar story to double prec.   THe advantage peaks at +50% efficiency at 1 thread/core.