@ahbarnett 2023-06-20
I am also seeing 20% speedups in 3D spreading on the xeon @ 1 thread/core, and 5-8% at OMP_NUM_THREADS=4.